laproteina_steerability/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "laproteina-steerability"
version = "0.1.0"
description = "Latent space diagnostics and property probes for La-Proteina"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "scipy",
    "scikit-learn",
    "pandas",
    "pyarrow",
    "matplotlib",
    "seaborn",
    "umap-learn",
    "pyyaml",
    "tqdm",
    "joblib",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
